# common: some optimizations in tracing code