--- a/src/common/cm/cm_test.cpp
+++ b/src/common/cm/cm_test.cpp
@@ -507,17 +507,12 @@ CM_BoundsIntersect
 */
 bool CM_BoundsIntersect( const vec3_t mins, const vec3_t maxs, const vec3_t mins2, const vec3_t maxs2 )
 {
-	if ( maxs[ 0 ] < mins2[ 0 ] - SURFACE_CLIP_EPSILON ||
-	     maxs[ 1 ] < mins2[ 1 ] - SURFACE_CLIP_EPSILON ||
-	     maxs[ 2 ] < mins2[ 2 ] - SURFACE_CLIP_EPSILON ||
-	     mins[ 0 ] > maxs2[ 0 ] + SURFACE_CLIP_EPSILON ||
-	     mins[ 1 ] > maxs2[ 1 ] + SURFACE_CLIP_EPSILON ||
-	     mins[ 2 ] > maxs2[ 2 ] + SURFACE_CLIP_EPSILON )
-	{
-		return false;
-	}
-
-	return true;
+	return ( maxs[ 0 ] >= mins2[ 0 ] - SURFACE_CLIP_EPSILON &&
+	     maxs[ 1 ] >= mins2[ 1 ] - SURFACE_CLIP_EPSILON &&
+	     maxs[ 2 ] >= mins2[ 2 ] - SURFACE_CLIP_EPSILON &&
+	     mins[ 0 ] <= maxs2[ 0 ] + SURFACE_CLIP_EPSILON &&
+	     mins[ 1 ] <= maxs2[ 1 ] + SURFACE_CLIP_EPSILON &&
+	     mins[ 2 ] <= maxs2[ 2 ] + SURFACE_CLIP_EPSILON );
 }
 
 /*
@@ -527,17 +522,12 @@ CM_BoundsIntersectPoint
 */
 bool CM_BoundsIntersectPoint( const vec3_t mins, const vec3_t maxs, const vec3_t point )
 {
-	if ( maxs[ 0 ] < point[ 0 ] - SURFACE_CLIP_EPSILON ||
-	     maxs[ 1 ] < point[ 1 ] - SURFACE_CLIP_EPSILON ||
-	     maxs[ 2 ] < point[ 2 ] - SURFACE_CLIP_EPSILON ||
-	     mins[ 0 ] > point[ 0 ] + SURFACE_CLIP_EPSILON ||
-	     mins[ 1 ] > point[ 1 ] + SURFACE_CLIP_EPSILON ||
-	     mins[ 2 ] > point[ 2 ] + SURFACE_CLIP_EPSILON )
-	{
-		return false;
-	}
-
-	return true;
+	return ( maxs[ 0 ] >= point[ 0 ] - SURFACE_CLIP_EPSILON &&
+	     maxs[ 1 ] >= point[ 1 ] - SURFACE_CLIP_EPSILON &&
+	     maxs[ 2 ] >= point[ 2 ] - SURFACE_CLIP_EPSILON &&
+	     mins[ 0 ] <= point[ 0 ] + SURFACE_CLIP_EPSILON &&
+	     mins[ 1 ] <= point[ 1 ] + SURFACE_CLIP_EPSILON &&
+	     mins[ 2 ] <= point[ 2 ] + SURFACE_CLIP_EPSILON );
 }
 
 // XreaL END