When spamming grenades, some parts of the tracing codes is iterated millions time. For example the `CM_TestInLeaf` function can count for 10 or 15% of the CPU time spent by the game, not because the compute is heavy, but because it is done so many time. In this case even simple things are costly, like dereferencing a pointer of pointer instead of deferencing a pointer, because it is done millions time.

So I turned some array index iterators into pointer iterators in those precious functions. While I was at it and for completeness, I also did it for every leaf brush, leaf surface, and brush side in both `tr_trace.cpp` and `tr_test.cpp`.

This PR also brings some commits unrolling some loops, helping the compiler to vectorize them, for similar reasons: small speedup can become large when done millions time.

I added an extra commit rewriting basically `if test; return true; else; return false;` into `return test`.