@DolceTriade 2023-03-31
Can we have a benchmark that shows the differences with these changes and without?

Any noticeable FPS increases? Any measurable CPU usage changes?