@illwieckz 2023-04-01
> Can we have a benchmark that shows the differences with these changes and without?
> 
> Any noticeable FPS increases? Any measurable CPU usage changes?

What I just did was to set my CPU in powersave mode and my GPU on high performance profile so the game spends more time in CPU code and the CPU time itself is less prone to precision loss (larger numbers). Then I loaded the chasm map on default spectator view (all snow flakes are particles doing collision tracing), waited 20s to be sure many particles were spawned, then recorded the game with Orbit for 20s. I did it with master branch then with this branch.

The given numbers are the percentage of CPU usage per function compared to the CPU usage of the whole game, in other words, where the game spends his time.

Function|Before|After
--|--|--
CM_TraceThroughLeaf|9.44%|7.96%
CM_Trace|5.56%|4.81%
CM_BoundsIntersect|4.42%|3.37%
CM_TransformedBoxTrace|3.64%|2.94%
CM_TraceThroughTree|2.57%|1.88%

What's also interesting is that in the mean time, other unrelated functions got their percentage increasing, which makes sense as the less percentage of time a given function takes in the whole time, the greater the percentage of unmodified time from unmodified function looks.

Function|Before|After
--|--|--
CG_AddPacketEntities|1.38%|1.61%
CG_GetLocation|1.21%|1.47%
CG_ClipMoveToEntities|0.76%|0.87%
CG_RenderParticles|0.72%|0.83%

Before:

[![trace iterate pointers](https://dl.illwieckz.net/b/unvanquished/wip/trace-iterate-pointers/20230401-193018-000.orbit-unvanquished.png)](https://dl.illwieckz.net/b/unvanquished/wip/trace-iterate-pointers/20230401-193018-000.orbit-unvanquished.png)

After:

[![trace iterate pointers](https://dl.illwieckz.net/b/unvanquished/wip/trace-iterate-pointers/20230401-193047-000.orbit-unvanquished.png)](https://dl.illwieckz.net/b/unvanquished/wip/trace-iterate-pointers/20230401-193047-000.orbit-unvanquished.png)