@illwieckz 2023-04-01
> Hmmm… there is a bug.
> 
> _Edit:_ that's in `cm_test: return boolean directly`.

I did a mistake while reversing the test, it's now fixed.